+++ app/src/main/java/com/techcourse/ManualHandlerMapping.java
@@ -3,12 +3,13 @@ package com.techcourse;
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.LoginViewController;
 import com.techcourse.controller.LogoutController;
+import com.techcourse.controller.RegisterController;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 import webmvc.org.springframework.web.servlet.mvc.asis.ForwardController;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handler.adapter.HandlerMapper;
+import webmvc.org.springframework.web.servlet.mvc.tobe.handler.mapper.HandlerMapper;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -24,8 +25,8 @@ public class ManualHandlerMapping implements HandlerMapper {
         controllers.put("/login", new LoginController());
         controllers.put("/login/view", new LoginViewController());
         controllers.put("/logout", new LogoutController());
-//        controllers.put("/register/view", new RegisterViewController());
-//        controllers.put("/register", new RegisterController());
+        controllers.put("/register", new RegisterController());
+//      controllers.put("/register/view", new RegisterViewController()); -> 어노테이션 기반 작동으로 변경
 
         log.info("Initialized Handler Mapping!");
         controllers.keySet()
