+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handler/mapper/AnnotationHandlerMapping.java
@@ -7,7 +7,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.bind.annotation.RequestMapping;
 import web.org.springframework.web.bind.annotation.RequestMethod;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handler.adapter.HandlerMapper;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
