+++ mvc/src/test/java/webmvc/org/springframework/web/servlet/mvc/tobe/handler/common/HandlerMappersTest.java
@@ -6,9 +6,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import samples.TestController;
-import webmvc.org.springframework.web.servlet.mvc.tobe.handler.adapter.HandlerMapper;
 import webmvc.org.springframework.web.servlet.mvc.tobe.handler.mapper.AnnotationHandlerMapping;
 import webmvc.org.springframework.web.servlet.mvc.tobe.handler.mapper.HandlerExecution;
+import webmvc.org.springframework.web.servlet.mvc.tobe.handler.mapper.HandlerMapper;
+import webmvc.org.springframework.web.servlet.mvc.tobe.handler.mapper.HandlerMappers;
 
 import java.lang.reflect.Method;
